Enable C4244 MS Compiler warning

+++ db/compaction_job.cc
@@ -409,9 +409,12 @@ void CompactionJob::GenSubcompactionBoundaries() {
 
   // Group the ranges into subcompactions
   const double min_file_fill_percent = 4.0 / 5;
-  uint64_t max_output_files = std::ceil(
-      sum / min_file_fill_percent /
-      cfd->GetCurrentMutableCFOptions()->MaxFileSizeForLevel(out_lvl));
+  uint64_t max_output_files = 
+    static_cast<uint64_t>(
+        std::ceil(
+          sum / min_file_fill_percent /
+          cfd->GetCurrentMutableCFOptions()->MaxFileSizeForLevel(out_lvl))
+          );
   uint64_t subcompactions =
       std::min({static_cast<uint64_t>(ranges.size()),
                 static_cast<uint64_t>(db_options_.max_subcompactions),
