+++ db/version_edit.h
@@ -55,7 +55,7 @@ struct FileDescriptor {
     return packed_number_and_path_id & kFileNumberMask;
   }
   uint32_t GetPathId() const {
-    return packed_number_and_path_id / (kFileNumberMask + 1);
+    return static_cast<uint32_t>(packed_number_and_path_id / (kFileNumberMask + 1));
   }
   uint64_t GetFileSize() const { return file_size; }
 };
